Moved DE440s fixture to a shareable location
